edge-worker-1/requirements.txt
# Core ML Frameworks
torch
torchvision # For loading CIFAR-10 data
numpy

# Communication Framework
grpcio
grpcio-tools

# Utility (for basic math/data handling)
pandas